Disable racy & flaky test (temporarily) (#33)
This was already done in the GitHub CI's make target (`test-unit`).

Adding the same skip in the `test-race` target too, which is used by the
full local invocation of `make`.

Not adding it in other targets (e.g. `test-system`) because they're not
used automatically so there's value in test failures there even now.