BUGFIX: pass context appropriately on callbacks to workflow (#91)
I introduced a bug in https://github.com/autokitteh/autokitteh/pull/89
where the context on a callback to workflow from runtime was not passed
correctly, thus callbacks to builtin functions was considered as a
nested activity, which cannot be supported. This compounded with another
bug in the same PR where builtins were not considered.

This PR also necessitated a change to the way I store starlark TLS data,
hence it's not very minimal, but still reasonable.