decouple deployments and sessions; sessions tests (#105)
- deployment and sessions are decoupled: while a deployment can be
specified for a session, it does not require it. session depends only on
build and env.
- change cli to accommodate for that
- add sessions test functionality a-la runtimes test. in the future will
be converted into unit tests (ENG-545).
- removed foreign key to deployments from sessions. need to add a
constraint instead (ENG-547).

btw, i would have loved to break this PR down a bit - but it's not too
bad. most of the lines are proto gen. also breaking it up is a bit
problematic in that there are changes in cmd that would have needed to
change again later, so isn't worth it.