http integration: return body functions (#137)
instead of returning `body_json`, `body_bytes` etc, which is... not
pretty, we return now functions that return the appropriate data. These
are functions and not just values since in the future we'll do it lazily
and not just duplicating the payload.

For that purpose, "const function" constructors are implemented that
just return their data as values.