Fix for sending form over HTTP without specifying content type (#193)
See lines 210-213 (not modified by this PR): if there's no custom
content-type, we set the default one in the HTTP request we're preparing
(`req.Header`), but not the helper variable (`contentType`). So in the
switch block (which evaluates at least a part of the content-type value)
we still need to consider that `contentType` may be `""`.