named integration ids to contain integration name hint (#190)
it is useful for integration ids that are built based on names to
include a hint for the name inside them. this change tries to embed the
original name in the id itself.

for example, for http:

`int_3kth00httpf1201a7ed83f7cd5`

this allows for easier debugging when need to know what integration this
id belongs to.