decouple cvars from env (#361)
Connection Vars should eventually be an implementation details of the
integration. This PR takes a step toward that goal by decoupling the env
module vars supplied to scripts from the actual connection vars by
adding a function to integrations to returns a configuration, instead of
reading directly from the connection vars.