feat(logger): configure min logging level by name, not ID (#532)
The default minimum logging level is
[Info](https://pkg.go.dev/go.uber.org/zap#pkg-constants). To change
this, set the following config value to the lower-case or all-caps name
of the level you want.

For example:

- CLI: `ak config set logger.level debug`

- Config YAML file:

  ```yaml
  logger:
    level: debug
  ```

- Environment variable: `AK_LOGGER__LEVEL=DEBUG`

Per @itayd and according to the principle of least surprise, invalid
names result in an error.

BREAKING CHANGE: Expected values are now strings instead of integers!
This PR still accepts old numeric values (e.g. -1 = debug), but this is
temporary and will be removed after all existing servers migrate.