feat(ENG-1274): add `required` to vars (#537)
Add an optional `required` field to vars. This allows to specify in the
manifest that a variable is required, and if a deploy attempt is made
without this variable being set, it will fail. This is helpful when we
want to specify required variables in manifest in kittehub, but without
giving any default value. This will force the user to explicitly set
these to deploy.

This also renames `is_secret` to `secret` for said vars, for
consistency.