fix: next_event can now process multiple sequential events (#546)
there was a problem where next_event skipped events. this change the
name of the last event read in sessions to be more explicit, as well as
the condition in the db to be more appropriate with the filter meaning.
also increase the seq number in the filter given to List as needed.