fix(ENG-1407): run next_event uuid generation in a side effect (#560)
so it looks like under some circumctances that involve signals, temporal
decides to rerun a workflow function even after it's completed in replay
mode.

when in replay, the next_event generates a new `signalID` though it
should have used the previous one. Properly putting it in a SideEffect
eliminates the problem.

also added an "optimization" to just not run the replay if the workflow
is already done.