fix(workflow): stop automaticaly delete drained deployments (#684)
This PR removes code that tries to delete draining deployment on last
session cleanup.

- There is a raise condition in deactivate_drain_deployment which is
running on every session cleanup. In cleanup it checks whether this is
the last session for deploymentID and whether the deployment is
draining. The problem/raise condition that deployment could be already
deleted by explicit call to project/deployment delete, thus it won't be
found and `deactivate_drain_deployment` activity will be stuck forever
(activities are not limited in time/attempts)

- Moreover running this code in session cleanup will lock sessions table
due to count(*) - when we are trying to check whether this is the last
session before deleting deployment.