feat: acceptance test for basic usability of integrations (#784)
This can be used for acceptance tests for new environments and auth
types, verification of connection UIs and flows, misc Python bug
reproductions, etc.

It is meant to be simpler and quicker than running all the samples one
after the other - a single deployment, trivial API calls without runtime
parameters, and no extra logic in triggers. In other words, is checks
that the rest of the projects are ready to be tested.

It's not meant to offer the same functionality coverage or educational
value as other projects in this repo.

FYI, some functionalities are currently broken in the cloud platform,
but are working in on-prem servers (e.g. #774 and #779 not deployed in
prod, and ENG-1680).