fix: `ak up -m test` get stuck sometimes (#914)
looks like the session tests get stuck because temporal start dev fails
to run from the binary. this attempts to start it multiple times on
failure. i tested it lots of times and now it looks to be solved using
this very heavy handed fix.

we should consider not starting temporal ourselves, but that might be
bothersome to the users in test mode.