fix: proper migration from 0.11.x versions to 0.12.x users and orgs (#928)
When the authz migration is done, we need to backfill proper user and
orgs configurations. For concrete example, all users now need to have a
default org id which points to a real org, and be a member of this org.
Also, all projects need to have an org associated with, which this
migration sets as the project's owner personal org.