fix(cron): fix invalid expiration times by renewing watches (#1022)
Oversight in the original implementation of Jira & Google workflows: if
a watch has an old-style/invalid timestamp, this is a reason to renew it
(to fix this problem), not to ignore it.

Specifically, this is what will facilitate a smooth transition for Jira
connections between pre-renewal Go-style timestamp strings (e.g.
`"2006-01-02 15:04:05.999999999 -0700 MST"`) and post-renewal RFC-3339
timestamp strings (e.g. `"2006-01-02T15:04:05Z07:00"`).