feat(sessions): get prints (#1040)
add another method to the sessions service: get prints. this will return
only the session prints. the new prints methods returns prints as
values, which for now will consist of only strings, but in the future
would be any arbitrary value (think js's `console.log`).

in the future the prints will be pulled out from the session log and
would be able to be retrieved only using get prints.