fix(zoom): oauth flow (#1056)
The original scopes I specified were legacy-style, which is no longer
allowed for new apps. Need to use only new granular scopes.

Anyway, passing an empty list is fine because it uses the apps default
scopes, and it's actually preferable because there are 100+ granular
scopes that we currently use.

Tested that an OAuth 3-legged flow really works, documented in
https://docs.autokitteh.com/integrations/zoom

Refs: INT-241