fix: high performance sqlite (#1007)
the instructions given in https://kerkour.com/sqlite-for-servers seem to
be working. tested under load, not a single BUSY was seen.

this essentially separates db connections into write and read
connections, limiting writes to a single connection. also add a bunch of
pragmas. for non-sqlite, there is no separation.

get rid of the sqlite lock.

also get rid of some unused functions.

Ref: ENG-190, ENG-648