refactor(CI): session tests (#1111)
Fix 1: `make test-sessions` didn't depend on `bin/ak` - this is a
mistake because it requires it, and we can't just assume that it's
up-to-date.

But... Fix 2: this means we also need to build AK and run `run.sh`
directly in the Go and Python CI workflows, to make GitHub caching
effective like in #1108, and avoid building AK twice.

Refs: INT-2014