fix: connstatus false positive (#1118)
Introduces a common function to validate that an OAuth access token
isn’t empty, confirming successful OAuth flow. A separate check was
added for Google due to differing variable names.
I haven’t tested the connStatus for every integration (lack of OAuth
apps), but since it worked on Slack, I assume it works for the rest.

refs: INT-225

---------

Co-authored-by: Daniel Abraham <daniel@autokitteh.com>