BUG-REPRO: python subscribe is stuck (#1129)
i modified the events system test to use python instead of starlark.
looks like when `subscribe` is called, python runner treats it as a
regular call and not the overridden call. this causes a deadlock inside
the runtime (it tries to send a syscall callback when it's stuck
processing the call in the select).

to repro:
```
$ make bin
$ go test -v ./tests/system -run 'TestSuite/sessions/events.txtar'
```