fix(salesforce): specify `eventType` and fix dispatch isolation bug (#1165)
**Considerations**: One option I considered for `eventType` was
combining the change type (which can be extracted from the sfdc event)
with the event type. For example, instead of `account`, it would be
`create_account`.

I decided against this because it would prevent users from receiving all
events related to a Salesforce entity. They can always filter for a
specific change type if needed.

**Testing**: Successfully received Salesforce events.  

_Refs: INT-321_