fix: make linear oauth actually use the actor var (#1295)
need to pass actor in the oauth URL. 

note that this change is made in a very specific way in order to change
as little as possible in the current flow. it is not clear to me why
auth0 is always privatized if it is customized. Refactoring this in
scope for ENG-2115.