feat: add deployment statistics table to db  (#1312)
Added `deployment_session_stats` table for the purpose of optimizing
`listDeploymentsWithStats`,
optimized the query by getting the count of the finished sessions from
the new table, instead of counting all sessions from the sessions table.

This will help in the long run for the enterprise db when the session
table is big.

ref: ENG-2142

---------

Co-authored-by: Efi Shtain <efi@autokitteh.com>