feat: values (#1320)
This adds a basic kv store to autokitteh scripts in starlark and python.
A new service in backend: "store", which exposes a Mutate and List
methods. Added to python sdk syscalls that call to the user_code
service, which in turn uses new Callbacks by the session to access the
store.

- Mutate(key, op, operands...): makes operation op on key using the
operands. Example: k1 set 1 will set k1 to 1.

Scope of store is project. Different projects cannot access each others
values.

TODO: We need to discuss how to limit access to these things.