feat: add AirTable integration (#1328)
- Added Airtable integration with `default OAuth` and `PAT` connectors
- Introduced PKCE handling in the `oauth` package

Note: @itayd — I'm not fully confident about how I handled the PKCE
logic in `getConfig`. The issue is that `getConfig` is called not only
during `startOAuthFlow`, but also in `exchangeCodeToToken` and in
`refreshOAuthToken`. This leads to a warning when `refreshOAuthToken`
tries to save a variable, which ideally shouldn't happen in that
context.
Open to suggestions or a better approach if you have one.

ref: ENG-434