fix: render structs as attr dicts in value to string (#1408)
will allow ui to properly render it.

This is, in principle, a theoretical breaking change for python users as
responses structures that come from activities are now AttrDicts and not
namedtuples, but this should these are virtually the same unless someone
tries to use int indexes on namedtuples, which there is no reason to.